Adding AlarmConfigTool and XmlModelWriter. Altering AlarmModelWriterTest.
AlarmConfigTool will be a command line tool for configuring the alarm
model by importing and exporting xml files.

XmlModelWriter contains the code to export the alarm model to an xml
file.

AlarmModelWriterTest altered to only contain the test code. It used to
contain the code that now resides in XmlModelWriter.