Refactored XmlModelWriter getModelXML method.
The method used to take an output stream as an argument however it only
used this stream once to initialize the writer member variable. This
initialization has been moved to the XmlModelWriter constructor. The
argument for an output stream has been removed from the getModelXML
method. The unit test was updated to use this new initialization scheme.
The code continued to pass the unit tests.