Implemented wait functionality for model stability.
The program now waits for the model to be stable for a specified time
before writing the model. Any update to the model will cause this time
to be reset, and the program then begins to wait again. Once the model
has been stable for the specified time, the program writes the model to
an xml file.