Removed 2 throttles. Added remove method to filter. Updated tests.
The area view now uses a single throttle to trigger additions, removals,
and updates of items in the area view.

When an item is removed from the area view, its entry in the area filter
should also be removed. A method was added to areaFilter to facilitate
this.

AreaFilterTest was updated to test the removeItem method.