Changed addPV and addComponent. Implemented rename on node.
To add a new PV or component, you now give the method the path to the
parent item instead of the parent itself. This is because there isn't a
guarantee you will have the parent item.

AlarmClient, AlarmConfigTool, AlarmTreeView, AddComponentAction, and
RenameComponentAction have all been changed to give the path to the
parent item.

RenameComponentAction has been implemented on node items as well as leaf
items. It will create a copy of the item and its subtree and the delete
the old item and its old subtree.