Fixed load past layout bug.
There was a bug where the final tab to be closed of the previous layout
would not be closed. This resulted in the new layout loaded from the
memento file to be accompanied by the remaining tab. This bug has been
fixed. Loading new layouts from memento files now results in only the
tabs described in the memento being loaded.