Altered validatePath, added tests.
validatePath now takes an AlarmTreeNode that should be the tree root
instead of the TreeView.
validatePath no longer checks for valid path elements as any string can
be a valid path element. AlarmTreeItem simply sets the name variable to
whatever string is given in the constructor. The path is constructed
from these strings. So don't check, they can be any string.