Adding voice test. Define voice dir in annunciator.
NotifyingConcurrentArrayList is a work in progress that should allow a
cleaner and better performing annunciator. Its a CopyOnWriteArrayList
that will execute a specific runnable on addition to and removal from
the internal array list. The goal is to allow for annunciation to occur
when the list is added to, as if the list had a listener.

In addition to this new list a timer should be all that is necessary to
implement an annunciator that doesn't depend on checking the
toAnnunciate list at a fixed rate.