Changed parseAlarmItemConfig to parseJsonText.
The method creates a parser for the passed json text and then returns
the root JsonNode of the parsed tree. This isn't specific to an
AlarmItemConfig. It is used with *Command Topic messages and *Talk topic
messages as well. Renaming it makes the code that uses it to parse
things other than alarm item configs more clear that the method isn't
specific to just alarm item configs.