No longer throwing exceptions when add/remove from selected lists.
The addSelected* and removeSelected* methods no longer throw exceptions
when an item is added or removed that is not in the available items
list. The methods now just return false to indicate they did not
complete the action.

The removal of this throwing has reduced the complexity of the list
selection view and the log books and tags view.