Implemented disable submit if required fields are not set.
The submit entry button in the LogEntryDialog is only enabled if
username, password, title, and selected logbooks are all not empty.
Should any of these fields be empty, their accompanying label's text
fill will be set to javafx.scene.paint.Color.RED. Once filled the label
returns to its default color.

After updating the log entry view to use LogService.createEntry()
a bug was introduced. Originally we requested a specific log factory and
ran the LogClient.createEntry method in a thread we create. However now,
since the log service uses an executor to create each entry in a
separate thread there is no guarantee as to when the entry will be
submitted. This led to temporary image files being deleted before the
submission occurred. This caused file not found exceptions. To work
around this, the toDelete list of temporary files and its associated
code has been commented out in the LogEntryModel.submitEntry() method.
The temporary image files are in the /tmp directory and are set to
delete on exit. But it would be good to find a way to delete them more
expediently.