Added save last entered username and password functionality.
Stores username and password into encrypted file upon submission. Reads
from same file upon start up to initialize the username and password
fields to the stored values. Reading from and writing to occurs on
background threads. Added synchronization to model's username and
password strings because they will be modified from the background
thread retrieving the stored values. Added a update credentials property
that views can listen to. This will be set when the credentials need to
be updated. The credential view listens to this and sets the username
and password field values when the property changes. Added getter
methods for username and password to the model. I didn't want to do this
originally as it allows confidential information to leak from the model.
However the view's that it would be leaking to already have the username
and password, so it doesn't seem to be an issue.