Scan server: Fix 'abort all'
When a running scan ends, it optionally updates the scan-active PV,
clearing it when all done, leaving it set when there are still other
idle or running scans.
On 'abort all', that running scan could find other idle scans in the
queue which are soon aborted, but by then the scan-active PV was already
set.

Change order of 'abort all', also first marking all scans to be aborted,
then aborting them, to keep the scan-active PV consistent.