Fall back to messages.properties on a per message level
Currently we only fall back to `messages.properties` if `messages_xx.properties` doesn't exist.
If we have an incomplete `messages_xx.properties`, we fall back to the name of the field in `Messages.java`.

This provides a worse experience than having no localization unless localizations are completely up to date.

With the changes in this commit, an incomplete localization will first fall back to english for the messages missing from `messages_xx.properties` and then to the field names for the messages missing from both `messages_xx.properties` and `messages.properties`.