Fixed 'Named font not being set' bug
When selecting a named font only the family,style,size combination was set but
not the name. You had to select another widget, come back to this widget (this
requirement is I believe fixed by the previous commit) and select the named
font again.
The cause of this bug is that the fontProperty listener changed the selection
of family, style, and size and the listeners for those called setFont()
with "incorrect" font settings (eg. family listener used the
not-yet-updated selection of style and size). The end result was the
correct font but it lost its name in the process. This commit fixes that
by not calling setFont() if setFont() is active.