LED: Fix erroneous initial value
When the PV happens to connect and send a value before the LED is first
rendered, it was initialized with `null` as a `new_value`, ignoring the
`runtimePropValue` that was already correct.
The LED would be stuck in that wrong state until the PV updates again.

This was observed in displays that show a PV in 'normal' widgets as well
as widgets that are dynamically added. For the latter widgets, the PV
was already connected, and the LED started out in the wrong state.