Fix logbook preference warning
`LogbookPreferences` has two settings:
`logbook_factory`, which is read from the preferences,
and `is_supported` which is automatically set based on the availability
of the selected logbook factory.

`logbook_factory` is correctly annotated as `@Preference`, since it's
initialized from preferences.

`is_supported` was erroneously annotated as `@Preference`, resulting in
this error on startup because there is no such setting in the preference
files, and there's not supposed to be one since the variable is
automatically set:
```
SEVERE [org.phoebus.framework.preferences] No default setting for
preference org.phoebus.logbook/is_supported
```