alarm server port
In 45e662959b97f31c4e84c7ec2cc44db9dc68b042, the example server was changed from `server=localhost:9092`  to `server=localhost:19092`.
The default port for running kafka out of the box still seems to be 9092, though. That's also what's mentioned in the alarm/Readme.md
Port 19092 seems to be popular in examples where Kafka is configured to allow additional connections, or when re-configuring it to control access.