Databrowser: Fix deleting traces with annotation.
Whhen deleting a trace with annotation, the annotation remained
orphaned. It was on the screen, but would no longer snap to any trace.
It would be saved in the *.plt file with -1 as the item index, and when
later loading the file, there'd be an error dialog.

This fixed the main issue by removing the annotation when deleting the
trace.

Note that there's no complete "undo": Undo will restore the trace, but
for now not recover the annotation for the trace. That would be another
thing to implement as time permits.