Tank: 'Horizontal' widget used space of 'vertical'
Because of the way the widget was rotated, it would capture mouse clicks
as if it was in the default, vertical orientation, ignoring the
transformation that makes it appear horizontally.
This might be a JavaFX error, but nevertheless prevented interacting
with widgets that are 'below' the horizontal tank.
Fixed by transforming the tank inside the pane instead of transforming
the pane.