Table: Changing data clears cell colors
Reason was a combination of script setting the same colors, so no
perceived change in coloring, and changes in data neither clearing nor
re-using the coloring.

Now changing the value will clear coloring, and then setting colors
actually has an effect.
Table example always contained a comment
"Changing the value of a table removes all cell colors."
so this was always the intended but not fully implemented behavior.