Split core-pv module (see #2939).
The core-pv module is split into separate modules, only leaving the core
classes and the internal PV sources in the original module. PV sources
external to Phoebus are separated, adding one module for each PV source
type.

The example Phoebus product now has explicit dependencies on all these
modules, so that it offers the same range of PV sources that it did
before.