Fix handling of multiple Kafka hosts.
When multiple Kafka hosts are specified as part of the alarm URI, the
URI cannot be parsed as a server-based URI, only as a registry-based
URI. This means that the getHosts(), getPort(), and getUserInfo()
methods return null. Therefore, we have to use getAuthority() instead
when cloning the URI.

This closes #3071.