ActionButton: Fix text (foreground) color
The text/foreground color was not dependably applied.

While we called `setTextFill(foreground)`, there was a contradicting
CSS entry that set the `.action_button_item .label` to `-fx-text-fill:
black`.
The widget would somewhat randomly show either the desired text color
or black as the display was edited.

To get the desired foreground (text) and background color in both the
plain action "button" and the "menu button" used when there are multiple
actions, setting the "-fx-text-fill" as well as the
"-fx-text-base-color" seems to be required, so all coloring is now based
on the widget style, not `setTextFill`.