Image widget: Fix VImage handling
When 'limits from PV' was used with a VImage (from PVA NTNDArray), the
image axes would update to the size of the image whenever the PV sent a
new value. But when the PV stayed constant and one would then perform
operations that require a layout, like show/hide the toolbar, the image
would be redrawn with the range from the model.
The display would thus show the wrong axes range until the PV sent
another update.

Now the model axes are updated from the VImage, so the display sticks to
the most recently received range from the VImage.