Fix label size so that it does not resize based on fontsize
If the fontsize was larger than the widget height then the Label
widget woult increase the height to fit the fontsize. This was
undesired behaviour and also not the behaviour of other widgets
displaying text.
Now we fix the label widget to be the size defined by the user
and the text will be cropped if too large for the widget.