Fix annotation moving to the left when X axis PV changes
When on plot the X axis PV changes the way that
there is no anymore the same x value, the annotation was choosing
the last index before.
When X axis PV is changing every time, it was causing the issue
that annotation was slowly going to the left and finally ending at the
beginning.
Finding the closest sample instead should fix the issue.